fix. 캐릭터, 카메라 값 수정
임시로 쓰던 파일을 사용할 파일로 변경.
