+++ Assets/Prefabs/Arrow/Bomb.prefab
@@ -109,6 +109,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d445fc62632d96f479025be83b900215, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  owner: {fileID: 0}
   countReturn: {fileID: 0}
   rb: {fileID: -2232580069837577341}
   damage: 0
@@ -188,7 +189,7 @@ SpriteRenderer:
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 0
   m_SortingLayer: 0
-  m_SortingOrder: 50
+  m_SortingOrder: 80
   m_Sprite: {fileID: 21300000, guid: 4c17f95ed34367b4e9e8913ec7b51751, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
