--- Assets/Scripts/Character.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86fc4b54b35d7c842a5fc97f97e6a9c4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
