+++ Assets/Scripts/Entity/Character.cs
@@ -4,78 +4,30 @@ using UnityEngine;
 
 public class Character : MonoBehaviour
 {
-    protected Rigidbody2D _rigidbody;
+    protected float maxHp = 100;
+    protected float curHp;
 
-    [SerializeField] private SpriteRenderer characterRenderer;
-    //[SerializeField] private Transform weaponPivot;
+	protected int arrowValue;
+    protected float moveSpeed;
+    protected float damage;
 
-    protected Vector2 movementDirection = Vector2.zero;
-    public Vector2 MovementDirection { get { return movementDirection; } }
+    protected float timer = 0.0f;
+    protected float attackDelay = 1.0f;
+    protected float arrowSpeed;
 
-    protected Vector2 lookDirection = Vector2.zero;
-    public Vector2 LookDirection { get { return lookDirection; } }
-
-    private Vector2 knockback = Vector2.zero;
-    private float knockbackDuration = 0.0f;
-
-    protected virtual void Awake()
+    protected Vector2 movementDirection;
+    protected Vector2 lookDirection;
+	protected Vector2 rot;
+    protected virtual void Init() 
     {
-        _rigidbody = GetComponent<Rigidbody2D>();
+        curHp = maxHp;
     }
 
-    protected virtual void Start()
+	protected virtual void Move() { }
+    protected virtual void Rotate() { }
+
+	protected virtual void Hit(ref float curHp, float damage) 
     {
-
-    }
-
-    protected virtual void Update()
-    {
-        HandleAction();
-        Rotate(lookDirection);
-    }
-
-    protected virtual void FixedUpdate()
-    {
-        Movment(movementDirection);
-        if (knockbackDuration > 0.0f)
-        {
-            knockbackDuration -= Time.fixedDeltaTime;
-        }
-    }
-
-    protected virtual void HandleAction()
-    {
-
-    }
-
-    private void Movment(Vector2 direction)
-    {
-        direction = direction * 5;
-        if (knockbackDuration > 0.0f)
-        {
-            direction *= 0.2f;
-            direction += knockback;
-        }
-
-        _rigidbody.velocity = direction;
-    }
-
-    private void Rotate(Vector2 direction)
-    {
-        float rotZ = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
-        bool isLeft = Mathf.Abs(rotZ) > 90f;
-
-        characterRenderer.flipX = isLeft;
-
-        //if (weaponPivot != null)
-        //{
-        //    weaponPivot.rotation = Quaternion.Euler(0, 0, rotZ);
-        //}
-    }
-
-    public void ApplyKnockback(Transform other, float power, float duration)
-    {
-        knockbackDuration = duration;
-        knockback = -(other.position - transform.position).normalized * power;
+        curHp -= damage;
     }
 }
