+++ Assets/Scripts/Entity/Character.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fb2032ce206098a4e8b52b978f507662
+guid: 86fc4b54b35d7c842a5fc97f97e6a9c4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
