+++ Assets/Scripts/Entity/FollowCamera.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5b3caeeaf7e1da643a438fd0ef974b23
+guid: 93fb2068c7868d040b26582bdc3373fc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
