+++ Assets/Scripts/Entity/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1f1d315e927268647878eabc73e00eb0
+guid: d5bef084f305d4f4fb3fdc338f6adaeb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
