--- Assets/Scripts/FollowCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 93fb2068c7868d040b26582bdc3373fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
