--- Assets/Scripts/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d5bef084f305d4f4fb3fdc338f6adaeb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
